fix: move higher-lower tests to separate file and sync docs to resolve CI failure
