+++ docs/contracts/higher-lower.md
@@ -90,48 +90,3 @@ pub fn get_game(env: Env, game_id: u64) -> Option<GameData>
 
 `Option<GameData>`
 
-### `set_result`
-```rust
-pub fn set_result(env: Env, game_id: u64, result: u32)
-```
-
-#### Parameters
-
-| Name | Type |
-|------|------|
-| `env` | `Env` |
-| `game_id` | `u64` |
-| `result` | `u32` |
-
-### `is_ready`
-```rust
-pub fn is_ready(env: Env, game_id: u64) -> bool
-```
-
-#### Parameters
-
-| Name | Type |
-|------|------|
-| `env` | `Env` |
-| `game_id` | `u64` |
-
-#### Return Type
-
-`bool`
-
-### `get_result`
-```rust
-pub fn get_result(env: Env, game_id: u64) -> u32
-```
-
-#### Parameters
-
-| Name | Type |
-|------|------|
-| `env` | `Env` |
-| `game_id` | `u64` |
-
-#### Return Type
-
-`u32`
-
